items + buttons

+++ bonus/src/player/utils.c
@@ -3,17 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   utils.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
+/*   By: escura <escura@student.42wolfsburg.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/19 11:57:44 by escura            #+#    #+#             */
-/*   Updated: 2024/08/12 11:53:34 by marvin           ###   ########.fr       */
+/*   Updated: 2024/08/24 16:47:40 by escura           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
 // Distance calculation function
-int distance(float x1, float y1, float x2, float y2)
+float distance(float x1, float y1, float x2, float y2)
 {
     return sqrtf((x2 - x1) * (x2 - x1) + (y2 - y1) * (y2 - y1));
 }
@@ -25,8 +25,8 @@ int view_lane_distance(float x1, float y1, float angle)
     
     float player_angle = player()->angle;
     
-    int raw_distance = distance(x1, y1, x2, y2);
-    int adjusted_distance = raw_distance * cos(player_angle - angle);
+    float raw_distance = distance(x1, y1, x2, y2);
+    float adjusted_distance = raw_distance * cos(player_angle - angle);
 
     return adjusted_distance;
 }
